Add additional Danger Types and Download Interrupts to Chromium (#29)
Based on my commit made in Feb
https://github.com/EricZimmerman/SQLECmd/pull/88 implements the same
missing values from the Chromium Source code. I noticed there is code in
Output/SQLiteHunter.yaml based on the definitions do I need to update
these too or is this sufficient? I haven't tested this on Velociraptor
but it should be fine as I followed the same format as the others.